extract outcome related helpers to src/helpers
